:sparkles: reaction picker awesomeness

+++ components/post/components/post-actions/components/PostReactButton.vue
@@ -1,5 +1,10 @@
 <template>
-    <v-popover class="has-width-100-percent ok-post-react-button-container" offset="20" placement="bottom-end">
+    <v-popover class="has-width-100-percent ok-post-react-button-container"
+               offset="20"
+               placement="bottom-end"
+               :open.sync="reactTooltipIsOpen"
+               trigger="manual"
+    >
         <!-- This will be the popover target (for the events and position) -->
         <button
                 class="button is-rounded ok-has-background-primary-highlight is-borderless is-fullwidth is-flex align-items-center"
@@ -16,7 +21,7 @@
                         </span>
         </button>
         <!-- This will be the content of the popover -->
-        <ok-reaction-picker slot="popover"/>
+        <ok-reaction-picker slot="popover" @onReactionPicked="onReactionEmojiPicked" :is-loading="requestInProgress"/>
     </v-popover>
 </template>
 
@@ -38,6 +43,8 @@
     import { TYPES } from "~/services/inversify-types";
     import { IUtilsService } from "~/services/utils-service/IUtilsService";
     import OkReactionPicker from "~/components/pickers/reaction-picker/ReactionPicker.vue";
+    import { IEmoji } from '~/models/common/emoji/IEmoji';
+    import { IPostReaction } from '~/models/posts/post-reaction/IPostReaction';
 
     @Component({
         name: "OkPostReactButton",
@@ -48,7 +55,7 @@
         private userService: IUserService = okunaContainer.get<IUserService>(TYPES.UserService);
         private utilsService: IUtilsService = okunaContainer.get<IUtilsService>(TYPES.UtilsService);
 
-
+        private reactWithEmojiOperation?: CancelableOperation<IPostReaction>;
         private removeReactionOperation?: CancelableOperation<void>;
 
         requestInProgress = false;
@@ -58,13 +65,15 @@
 
         destroyed() {
             if (this.removeReactionOperation) this.removeReactionOperation.cancel();
+            if (this.reactWithEmojiOperation) this.reactWithEmojiOperation.cancel();
         }
 
+        get postHasReaction(){
+            return !! this.post.reaction;
+        }
 
         async onReactButtonPressed() {
             if (this.requestInProgress) return;
-
-            this.requestInProgress = true;
 
             if (this.post.reaction) {
                 await this.removeCurrentReaction();
@@ -72,7 +81,10 @@
                 this.openReactionsPicker();
             }
 
-            this.requestInProgress = false;
+        }
+
+        onReactionEmojiPicked(emoji: IEmoji){
+            this.reactWithEmoji(emoji);
         }
 
         private openReactionsPicker() {
@@ -80,6 +92,8 @@
         }
 
         private async removeCurrentReaction() {
+            if(this.requestInProgress) return;
+            this.requestInProgress = true;
             try {
                 this.removeReactionOperation = CancelableOperation.fromPromise(this.userService.deletePostReaction({
                     post: this.post,
@@ -90,6 +104,29 @@
                 this.utilsService.handleErrorWithToast(error);
             } finally {
                 this.removeReactionOperation = null;
+                this.requestInProgress = false;
+            }
+        }
+
+
+        private async reactWithEmoji(emoji: IEmoji) {
+            if(this.requestInProgress) return;
+            this.requestInProgress = true;
+
+            try {
+                this.reactWithEmojiOperation = CancelableOperation.fromPromise(this.userService.reactToPost({
+                    post: this.post,
+                    emoji: emoji
+                }));
+
+                const postReaction = await this.reactWithEmojiOperation.value;
+                this.post.setReaction(postReaction);
+                this.reactTooltipIsOpen = false;
+            } catch (error) {
+                this.utilsService.handleErrorWithToast(error);
+            } finally {
+                this.reactWithEmojiOperation = null;
+                this.requestInProgress = false;
             }
         }
